First mods based on discussions (#368)
